refactor: Result 도메인 객체로 취급하지 않고 출력용으로 이용하도록 수정, 기존 메서드는 Tickets 객체로 이동
