+++ src/test/java/lotto/domain/TicketsTest.java
@@ -23,4 +23,21 @@ class TicketsTest {
 
         assertThat(countPerPrize.toString()).isEqualTo(tickets.makeStatistics(winningNumbers).toString());
     }
+
+    @Test
+    void test_당첨금계산() {
+        Numbers winningNumbers = new Numbers("1, 2, 3, 4, 5, 6");
+        Tickets tickets = new Tickets(List.of(new Ticket(new Numbers("1, 2, 3, 4, 5, 6")), new Ticket(new Numbers("1, 2, 3, 4, 5, 45"))));
+
+        assertThat(tickets.calcWinningAmount(winningNumbers)).isEqualTo(2000000000L + 150000000L);
+    }
+
+    @Test
+    void test_수익률계산() {
+        Numbers winningNumbers = new Numbers("1, 2, 3, 4, 5, 6");
+        Tickets tickets = new Tickets(List.of(new Ticket(new Numbers("1, 2, 3, 4, 5, 6")), new Ticket(new Numbers("1, 2, 3, 4, 5, 45"))));
+
+        double rateOfReturn = (double) (2000000000L + 150000000L) / 2000L;
+        assertThat(tickets.calcRateOfReturn(winningNumbers)).isEqualTo(rateOfReturn);
+    }
 }
